be/requirements.txt
# Core FastAPI dependencies
fastapi[standard]==0.115.8
pydantic==2.10.6
starlette==0.45.3
uvicorn==0.34.0

# Additional runtime dependencies
python-dotenv==1.0.1
python-multipart==0.0.20
email_validator==2.2.0
pydantic-core==2.27.2
typing-extensions==4.12.2

# Server dependencies
uvloop==0.21.0
httptools==0.6.4
watchfiles==1.0.4
websockets==15.0
requests==2.32.3
urllib3-2.3.0

# HTTP client
httpx==0.28.1
requests == 2.26.0

# OpenAI and OCR dependencies
openai==1.65.1
pytesseract==0.3.10
Pillow==11.1.0
pdfplumber==0.10.3
pymupdf==1.24.10
pymupdf4llm==0.0.17
numpy==1.21.0
